refactor(codec): own operation options
