+++ crates/copybook-record-io/tests/comprehensive_dispatch_tests.rs
@@ -4,8 +4,8 @@
 //! Comprehensive integration tests for record-io dispatch covering
 //! fixed-length reading, RDW reading, error handling, and edge cases.
 
+use copybook_codec::options::RecordFormat;
 use copybook_error::ErrorCode;
-use copybook_options::RecordFormat;
 use copybook_record_io::{read_record, write_record};
 use std::io::{self, Cursor, Read, Write};
